Fixes made on the CTE of the tables, with casting from smallint to integer, of relevant columns. Updated schema with tests and descriptions.

+++ dvd_rental_proj/models/staging/stg__dvd_rental_staff.sql
@@ -5,8 +5,8 @@ WITH staff_source AS (
 SELECT staff_id,
 first_name,
 last_name,
-address_id,
-store_id,
+address_id::INTEGER,
+store_id::INTEGER,
 active,
 username,
 last_update
